requirements.txt
numpy
matplotlib
scipy
noise
einops
python-dotenv
openai
ipython